+++ .sequelizerc
@@ -1,8 +1,8 @@
 const path = require('path')
 
 module.exports = {
-  'config': path.resolve('src', 'db', 'config', 'config.js'),
-  'models-path': path.resolve('src', 'db', 'models'),
-  'seeders-path': path.resolve('src', 'db', 'seeders'),
-  'migrations-path': path.resolve('src', 'db', 'migrations'),
+  'config': path.resolve('sequelize', 'config', 'config.js'),
+  'models-path': path.resolve('src', 'models'),
+  'seeders-path': path.resolve('sequelize', 'seeders'),
+  'migrations-path': path.resolve('sequelize', 'migrations'),
 }
