+++ package.json
@@ -32,8 +32,8 @@
     "pg": "^8.2.1",
     "pg-hstore": "^2.3.3",
     "regenerator-runtime": "^0.13.5",
-    "sequelize": "^5.21.13",
-    "sequelize-cli": "^5.5.1",
+    "sequelize": "^6.3.0",
+    "sequelize-cli": "^6.2.0",
     "winston": "^3.2.1",
     "yup": "^0.29.1"
   },
