+++ src/api/auth/auth-controller.js
@@ -1,5 +1,5 @@
 import * as yup from 'yup'
-import Database from '../../db/models'
+import Database from '../../models'
 
 const { User } = Database
